4.4. Максимум из трех чисел [#190][#55787]

+++ chapter_001/src/test/java/ru/job4j/max/MaxTest.java
@@ -1,6 +1,5 @@
 package ru.job4j.max;
 import org.junit.Test;
-import ru.job4j.max.Max;
 import static  org.hamcrest.core.Is.is;
 import static org.junit.Assert.assertThat;
 
@@ -9,7 +8,7 @@ public class MaxTest {
     @Test
     public void whenFirstLessSecond() {
         Max maks = new Max();
-        int result = maks.maxim(3, 1, 8);
+        double result = maks.maxim(3, 1, 8);
         assertThat(result, is(8));
     }
